+++ crates/grafbase-sdk/wit/since_0_14_0/authentication.wit
@@ -1,0 +1,9 @@
+interface authentication {
+    use headers.{headers};
+    use error.{error-response};
+    use token.{token};
+
+    authenticate: func(
+        headers: headers,
+    ) -> result<token, error-response>;
+}
