+++ crates/grafbase-sdk/wit/since_0_14_0/field-resolver-types.wit
@@ -1,0 +1,8 @@
+interface field-resolver-types {
+    use error.{error};
+
+    record field-output {
+        // list in the same order as input.edges with data serialized in CBOR
+        outputs: list<result<list<u8>, error>>
+    }
+}
