+++ crates/grafbase-sdk/wit/since_0_14_0/selection-set-resolver.wit
@@ -1,0 +1,3 @@
+interface selection-set-resolver {
+
+}
