+++ crates/grafbase-sdk/wit/since_0_14_0/world.wit
@@ -1,0 +1,33 @@
+package grafbase:sdk;
+
+world sdk {
+    import access-log;
+    import cache;
+    import error;
+    import headers;
+    import http-client;
+    import nats-client;
+    import token;
+    import directive;
+    import schema;
+    import authorization-types;
+    import field-resolver-types;
+
+    export authentication;
+    export authorization;
+    export field-resolver;
+    export selection-set-resolver;
+
+    use directive.{schema-directive};
+
+    // The extension registration function. Must be called before initialization.
+    export register-extension: func();
+
+    // initialization function called to set up the wasm extension
+    // if an error happens here, the gateway will refuse to continue.
+    // Receives a list of schema directives associated with the extension
+    export init: func(
+        schema-directives: list<schema-directive>,
+        configuration: list<u8>,
+    ) -> result<_, string>;
+}
