+++ crates/integration-tests/data/extensions/crates/authorization-010/Cargo.toml
@@ -8,7 +8,7 @@ keywords.workspace = true
 repository.workspace = true
 
 [dependencies]
-grafbase-sdk.workspace = true
+grafbase-sdk = { version = "0.13.0" }
 http.workspace = true
 rkyv.workspace = true
 serde.workspace = true
