+++ crates/wasi-component-loader/src/extension/api/mod.rs
@@ -1,11 +1,13 @@
 mod since_0_10_0;
+mod since_0_14_0;
 mod since_0_8_0;
 mod since_0_9_0;
 
 use since_0_8_0::instance::ExtensionInstanceSince080;
 use since_0_9_0::instance::ExtensionInstanceSince090;
 use since_0_10_0::ExtensionInstanceSince0_10_0;
-pub use since_0_10_0::world as wit;
+use since_0_14_0::ExtensionInstanceSince0_14_0;
+pub use since_0_14_0::world as wit;
 
 use super::ExtensionInstance;
 use crate::WasiState;
@@ -19,6 +21,7 @@ pub(crate) enum SdkPre {
     Since0_8_0(since_0_8_0::wit::SdkPre<WasiState>),
     Since0_9_0(since_0_9_0::wit::SdkPre<WasiState>),
     Since0_10_0(since_0_10_0::SdkPre<WasiState>),
+    Since0_14_0(since_0_14_0::wit::SdkPre<WasiState>),
 }
 
 impl SdkPre {
@@ -40,10 +43,15 @@ impl SdkPre {
                 let instance_pre = linker.instantiate_pre(&component)?;
                 SdkPre::Since0_9_0(since_0_9_0::wit::SdkPre::<WasiState>::new(instance_pre)?)
             }
-            _ => {
+            v if v < &Version::new(0, 14, 0) => {
                 since_0_10_0::Sdk::add_to_linker(&mut linker, |state| state)?;
                 let instance_pre = linker.instantiate_pre(&component)?;
                 SdkPre::Since0_10_0(since_0_10_0::SdkPre::<WasiState>::new(instance_pre)?)
+            }
+            _ => {
+                since_0_14_0::Sdk::add_to_linker(&mut linker, |state| state)?;
+                let instance_pre = linker.instantiate_pre(&component)?;
+                SdkPre::Since0_14_0(since_0_14_0::SdkPre::<WasiState>::new(instance_pre)?)
             }
         })
     }
@@ -105,6 +113,21 @@ impl SdkPre {
                     .await??;
 
                 let instance = ExtensionInstanceSince0_10_0 {
+                    store,
+                    inner,
+                    poisoned: false,
+                };
+
+                Ok(Box::new(instance))
+            }
+            SdkPre::Since0_14_0(sdk_pre) => {
+                let mut store = Store::new(sdk_pre.engine(), state);
+
+                let inner = sdk_pre.instantiate_async(&mut store).await?;
+                inner.call_register_extension(&mut store).await?;
+                inner.call_init(&mut store, schema_directives, guest_config).await??;
+
+                let instance = ExtensionInstanceSince0_14_0 {
                     store,
                     inner,
                     poisoned: false,
