+++ crates/wasi-component-loader/src/extension/api/since_0_10_0/instance/selection_set_resolver.rs
@@ -1,0 +1,3 @@
+use crate::extension::SelectionSetResolverExtensionInstance;
+
+impl SelectionSetResolverExtensionInstance for super::ExtensionInstanceSince0_10_0 {}
