+++ crates/wasi-component-loader/src/extension/api/since_0_14_0/mod.rs
@@ -1,7 +1,7 @@
 mod instance;
 pub mod wit;
 
-pub use instance::ExtensionInstanceSince0_10_0;
+pub use instance::ExtensionInstanceSince0_14_0;
 pub use wit::{Sdk, SdkPre};
 
 #[allow(unused)]
@@ -9,9 +9,8 @@ pub mod world {
     use super::wit::exports::grafbase::sdk as exports;
     use super::wit::grafbase::sdk;
 
-    pub use exports::authorization::{AuthorizationDecisions, AuthorizationDecisionsDenySome};
-    pub use exports::resolver::FieldOutput;
     pub use sdk::access_log::{AccessLog, LogError};
+    pub use sdk::authorization_types::{AuthorizationDecisions, AuthorizationDecisionsDenySome};
     pub use sdk::cache::Cache;
     pub use sdk::directive::{
         DirectiveSite, EnumDirectiveSite, FieldDefinitionDirective, FieldDefinitionDirectiveSite,
@@ -19,6 +18,7 @@ pub mod world {
         ScalarDirectiveSite, SchemaDirective, UnionDirectiveSite,
     };
     pub use sdk::error::{Error, ErrorResponse};
+    pub use sdk::field_resolver_types::FieldOutput;
     pub use sdk::headers::{HeaderError, Headers};
     pub use sdk::http_client::{HttpClient, HttpError, HttpMethod, HttpRequest, HttpResponse};
     pub use sdk::nats_client::{NatsAuth, NatsKeyValue, NatsStreamConfig, NatsStreamDeliverPolicy, NatsSubscriber};
