+++ crates/wasi-component-loader/src/extension/api/since_0_14_0/wit/field_resolver_types.rs
@@ -1,0 +1,26 @@
+use engine_error::{ErrorCode, GraphqlError};
+use runtime::extension::Data;
+
+use crate::state::WasiState;
+
+pub use super::grafbase::sdk::field_resolver_types::*;
+
+impl Host for WasiState {}
+
+impl From<FieldOutput> for Vec<Result<Data, GraphqlError>> {
+    fn from(value: FieldOutput) -> Self {
+        let mut results = Vec::new();
+
+        for result in value.outputs {
+            match result {
+                Ok(data) => results.push(Ok(Data::CborBytes(data))),
+                Err(error) => {
+                    let error = error.into_graphql_error(ErrorCode::InternalServerError);
+                    results.push(Err(error))
+                }
+            }
+        }
+
+        results
+    }
+}
