+++ crates/wasi-component-loader/src/extension/api/since_0_14_0/wit/schema.rs
@@ -1,0 +1,5 @@
+use crate::state::WasiState;
+
+pub use super::grafbase::sdk::schema::*;
+
+impl Host for WasiState {}
