+++ crates/wasi-component-loader/src/extension/api/since_0_8_0/instance/selection_set_resolver.rs
@@ -1,0 +1,3 @@
+use crate::extension::SelectionSetResolverExtensionInstance;
+
+impl SelectionSetResolverExtensionInstance for super::ExtensionInstanceSince080 {}
