+++ crates/wasi-component-loader/src/extension/manager/instance/selection_set_resolver.rs
@@ -1,0 +1,1 @@
+pub(crate) trait SelectionSetResolverExtensionInstance {}
